requirements.txt
accelerate==1.8.1
acres==0.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
banks==2.1.3
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
chromadb==0.5.0
ci-info==0.3.0
click==8.2.1
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
dataclasses-json==0.6.7
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
docx2txt==0.9
etelemetry==0.3.1
filelock==3.18.0
filetype==1.2.0
fitz==0.0.1.dev2
frozenlist==1.7.0
fsspec==2025.5.1
greenlet==3.2.3
griffe==1.7.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.33.2
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
llama-cloud==0.1.30
llama-cloud-services==0.6.41
llama-index==0.12.47
llama-index-agent-openai==0.4.12
llama-index-cli==0.4.4
llama-index-core==0.12.47
llama-index-embeddings-huggingface==0.5.5
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.7.8
llama-index-instrumentation==0.2.0
llama-index-llms-huggingface==0.5.0
llama-index-llms-openai==0.4.7
llama-index-multi-modal-llms-openai==0.5.1
llama-index-program-openai==0.3.2
llama-index-question-gen-openai==0.3.1
llama-index-readers-file==0.4.9
llama-index-readers-llama-parse==0.4.0
llama-index-workflows==1.0.1
llama-parse==0.6.41
looseversion==1.3.0
lxml==6.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
nltk==3.9.1
numpy==2.3.1
openai==1.93.0
packaging==25.0
pandas==2.2.3
pathlib==1.0.1
pillow==11.3.0
platformdirs==4.3.8
propcache==0.3.2
prov==2.1.1
psutil==7.0.0
puremagic==1.30
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
PyMuPDF==1.26.3
pyparsing==3.2.3
pypdf==5.7.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==7.1.4
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
sentence-transformers==5.0.0
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
striprtf==0.0.26
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
traits==7.0.2
transformers==4.53.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wrapt==1.17.2
yarl==1.20.1
